+++ Client/Frontend/TabContentsScripts/FindInPageHelper.swift
