+++ Client/Frontend/TabContentsScripts/LoginsHelper.swift
