+++ Client/Frontend/TabContentsScripts/NightModeHelper.swift
