+++ Client/Frontend/TabContentsScripts/NoImageModeHelper.swift
