+++ Client/Frontend/TabContentsScripts/SessionRestoreHelper.swift
