+++ Client/Frontend/TabContentsScripts/UserScriptManager.swift
