+++ Client/Frontend/Toolbar+URLBar/ToolbarButton.swift
@@ -1,0 +1,60 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import UIKit
+
+class ToolbarButton: UIButton {
+    var selectedTintColor: UIColor!
+    var unselectedTintColor: UIColor!
+    var disabledTintColor = UIColor.Photon.Grey50
+
+    // Optionally can associate a separator line that hide/shows along with the button
+    weak var separatorLine: UIView?
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        adjustsImageWhenHighlighted = false
+        selectedTintColor = tintColor
+        unselectedTintColor = tintColor
+        imageView?.contentMode = .scaleAspectFit
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override open var isHighlighted: Bool {
+        didSet {
+            self.tintColor = isHighlighted ? selectedTintColor : unselectedTintColor
+        }
+    }
+
+    override open var isEnabled: Bool {
+        didSet {
+            self.tintColor = isEnabled ? unselectedTintColor : disabledTintColor
+        }
+    }
+
+    override var tintColor: UIColor! {
+        didSet {
+            self.imageView?.tintColor = self.tintColor
+        }
+    }
+
+    override var isHidden: Bool {
+        didSet {
+            separatorLine?.isHidden = isHidden
+        }
+    }
+}
+
+extension ToolbarButton: NotificationThemeable {
+    func applyTheme() {
+        selectedTintColor = UIColor.theme.toolbarButton.selectedTint
+        disabledTintColor = UIColor.theme.toolbarButton.disabledTint
+        unselectedTintColor = UIColor.theme.browser.tint
+        tintColor = isEnabled ? unselectedTintColor : disabledTintColor
+        imageView?.tintColor = tintColor
+    }
+}
