+++ Client/Frontend/Toolbar+URLBar/URLTextField.swift
@@ -1,0 +1,27 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import UIKit
+import Shared
+
+class URLTextField: UITextField {
+    weak var accessibilityActionsSource: AccessibilityActionsSource?
+
+    override var accessibilityCustomActions: [UIAccessibilityCustomAction]? {
+        get {
+            return accessibilityActionsSource?.accessibilityCustomActionsForView(self)
+        }
+        set {
+            super.accessibilityCustomActions = newValue
+        }
+    }
+
+    override var canBecomeFirstResponder: Bool {
+        return false
+    }
+
+    override func textRect(forBounds bounds: CGRect) -> CGRect {
+        return bounds.insetBy(dx: TabLocationViewUX.Spacing, dy: 0)
+    }
+}
